lettaaugment-source/requirements.txt
# API Server
flask==3.0.0
werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.3
httpx==0.28.1
colorama==0.4.6
openai==1.57.3

# Weaviate
weaviate-client==4.13.2

# Utilities
pydantic==2.11.3
typing-extensions==4.13.2
schedule==1.2.1
pytz==2024.1

# Async support
aiosignal==1.3.1
attrs==25.3.0
yarl==1.9.4  # Updated for Python 3.12 compatibility
multidict==6.4.3
frozenlist==1.5.0
async-timeout==4.0.3
aiofiles==24.1.0

# ASGI server support
asgiref==3.7.2
hypercorn==0.15.0
quart==0.19.4
